requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2==2.9.10
requests
sqlalchemy
python-telegram-bot